+++ Apps/Examples/Examples/Models/Examples.swift
@@ -149,6 +149,8 @@ struct Examples {
                 type: CombineLocationExample.self)
         }
     }
+    
+    public static let geofencingExample = Example(title: "Geofencing", description: "Showcase geofencing detection", type: GeofencingExample.self)
 
     // Examples that focus on displaying the user's location.
     public static let locationExamples: [Example] = .init {
@@ -167,7 +169,9 @@ struct Examples {
         Example(title: "Simulate navigation",
                 description: "Simulate a driving trip from LA to San Francisco along a pre-defined route.",
                 type: NavigationSimulatorExample.self)
+        geofencingExample
     }
+
 
     // Examples that highlight using the Offline APIs.
     static let offlineExamples: [Example] = .init {